+++ tests/testthat/test-geo_amenity.R
@@ -31,4 +31,10 @@ test_that("Checking query", {
     "pub",
     custom_query = list(extratags = 1)
   )), 1)
+  expect_lt(nrow(geo_amenity(
+    bbox = c(-1.1446, 41.5022, -0.4854, 41.8795),
+    "pub",
+    limit = 1,
+    strict = TRUE
+  )), 2)
 })